requirements.txt
langchain
langgraph
groq
pandas
matplotlib
seaborn
plotly
kaggle
python-dotenv
